handle roles and permissions in dashboards

+++ app/Http/Controllers/Backend/Dashboards/Admin/AnnouncementController.php
@@ -20,16 +20,22 @@ class AnnouncementController extends Controller
 
     public function index()
     {
+        // apply permissions
+        abort_if(!hasPermission('view announcements'), 403, __('You are not authorized to view announcements'));
         return view('backend.dashboards.admin.pages.announcements.index');
     }
 
     public function data()
     {
+        // apply permissions
+        abort_if(!hasPermission('view announcements'), 403, __('You are not authorized to view announcements'));
         return $this->repo->data();
     }
 
     public function create()
     {
+        // apply permissions
+        abort_if(!hasPermission('create announcement'), 403, __('You are not authorized to create announcement'));
         $clinics = Clinic::select('id','name')->orderBy('name')->get();
         $suppliers = Supplier::select('id','name')->orderBy('name')->get();
         return view('backend.dashboards.admin.pages.announcements.create', compact('clinics','suppliers'));
@@ -37,11 +43,15 @@ class AnnouncementController extends Controller
 
     public function store(StoreAnnouncementRequest $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('create announcement'), 403, __('You are not authorized to create announcement'));
         return $this->repo->store($request);
     }
 
     public function edit($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update announcement'), 403, __('You are not authorized to view announcement'));
         $announcement = $this->repo->show($id);
         $clinics = Clinic::select('id','name')->orderBy('name')->get();
         $suppliers = Supplier::select('id','name')->orderBy('name')->get();
@@ -50,12 +60,15 @@ class AnnouncementController extends Controller
 
     public function update(UpdateAnnouncementRequest $request, $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update announcement'), 403, __('You are not authorized to update announcement'));
         return $this->repo->update($request, $id);
     }
 
     public function destroy($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('delete announcement'), 403, __('You are not authorized to delete announcement'));
         return $this->repo->destroy($id);
     }
 }
-
