+++ app/Http/Controllers/Backend/Dashboards/Admin/BlogPostController.php
@@ -24,11 +24,15 @@ class BlogPostController extends Controller
      */
     public function index()
     {
+        // apply permissions
+        abort_if(!hasPermission('view blog posts'), 403, __('You are not authorized to view blog posts'));
         return view('backend.dashboards.admin.pages.blog-posts.index');
     }
 
     public function data()
     {
+        // apply permissions
+        abort_if(!hasPermission('view blog posts'), 403, __('You are not authorized to view blog posts'));
         return $this->blogPostRepo->data();
     }
 
@@ -37,6 +41,8 @@ class BlogPostController extends Controller
      */
     public function create()
     {
+        // apply permissions
+        abort_if(!hasPermission('create blog post'), 403, __('You are not authorized to create blog post'));
         $blogCategories = BlogCategory::select('id', 'name_en')->get();
         return view('backend.dashboards.admin.pages.blog-posts.create', compact('blogCategories'));
     }
@@ -46,6 +52,8 @@ class BlogPostController extends Controller
      */
     public function store(StoreBlogPostRequest $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('create blog post'), 403, __('You are not authorized to create blog post'));
         return $this->blogPostRepo->store($request);
     }
 
@@ -54,6 +62,8 @@ class BlogPostController extends Controller
      */
     public function show($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('view blog posts'), 403, __('You are not authorized to view blog post'));
         $blogPost = $this->blogPostRepo->show($id);
 
         return request()->ajax()
@@ -66,6 +76,8 @@ class BlogPostController extends Controller
      */
     public function edit($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update blog post'), 403, __('You are not authorized to update blog post'));
         $blogPost = $this->blogPostRepo->show($id);
         $blogCategories = BlogCategory::select('id', 'name_en')->get();
 
@@ -77,11 +89,15 @@ class BlogPostController extends Controller
      */
     public function update(UpdateBlogPostRequest $request, $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update blog post'), 403, __('You are not authorized to update blog post'));
         return $this->blogPostRepo->update($request, $id);
     }
 
     public function updateStatus(Request $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('update blog post status'), 403, __('You are not authorized to update blog post status'));
         return $this->blogPostRepo->updateStatus($request);
     }
 
@@ -90,26 +106,36 @@ class BlogPostController extends Controller
      */
     public function destroy($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('delete blog post'), 403, __('You are not authorized to delete blog post'));
         return $this->blogPostRepo->destroy($id);
     }
 
     public function trash()
     {
+        // apply permissions
+        abort_if(!hasPermission('view trash blog posts'), 403, __('You are not authorized to view trash blog posts'));
         return view('backend.dashboards.admin.pages.blog-posts.trash');
     }
 
     public function trashData()
     {
+        // apply permissions
+        abort_if(!hasPermission('view trash blog posts'), 403, __('You are not authorized to view trash blog posts'));
         return $this->blogPostRepo->trashData();
     }
 
     public function restore($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('restore blog post'), 403, __('You are not authorized to restore blog post'));
         return $this->blogPostRepo->restore($id);
     }
 
     public function forceDelete($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('force delete blog post'), 403, __('You are not authorized to force delete blog post'));
         return $this->blogPostRepo->forceDelete($id);
     }
 }
