+++ app/Http/Controllers/Backend/Dashboards/Admin/JobController.php
@@ -23,11 +23,15 @@ class JobController extends Controller
      */
     public function index()
     {
+        // apply permissions
+        abort_if(!hasPermission('view jobs'), 403, __('You are not authorized to view jobs'));
         return view('backend.dashboards.admin.pages.jobs.index');
     }
 
     public function data()
     {
+        // apply permissions
+        abort_if(!hasPermission('view jobs'), 403, __('You are not authorized to view jobs'));
         return $this->jobRepo->data();
     }
 
@@ -36,6 +40,8 @@ class JobController extends Controller
      */
     public function create()
     {
+        // apply permissions
+        abort_if(!hasPermission('create job'), 403, __('You are not authorized to create job'));
         return view('backend.dashboards.admin.pages.jobs.create');
     }
 
@@ -44,6 +50,8 @@ class JobController extends Controller
      */
     public function store(StoreJobRequest $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('create job'), 403, __('You are not authorized to create job'));
         return $this->jobRepo->store($request);
     }
 
@@ -52,6 +60,8 @@ class JobController extends Controller
      */
     public function show($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('view job'), 403, __('You are not authorized to view job'));
         $job = $this->jobRepo->show($id);
 
         return request()->ajax()
@@ -64,6 +74,8 @@ class JobController extends Controller
      */
     public function edit($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update job'), 403, __('You are not authorized to update job'));
         $job = $this->jobRepo->show($id);
 
         return view('backend.dashboards.admin.pages.jobs.edit', compact('job'));
@@ -74,11 +86,15 @@ class JobController extends Controller
      */
     public function update(UpdateJobRequest $request, $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update job'), 403, __('You are not authorized to update job'));
         return $this->jobRepo->update($request, $id);
     }
 
     public function updateStatus(Request $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('toggle job status'), 403, __('You are not authorized to update job status'));
         return $this->jobRepo->updateStatus($request);
     }
 
@@ -87,26 +103,36 @@ class JobController extends Controller
      */
     public function destroy($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('delete job'), 403, __('You are not authorized to delete job'));
         return $this->jobRepo->destroy($id);
     }
 
     public function trash()
     {
+        // apply permissions
+        abort_if(!hasPermission('view trash jobs'), 403, __('You are not authorized to view trash jobs'));
         return view('backend.dashboards.admin.pages.jobs.trash');
     }
 
     public function trashData()
     {
+        // apply permissions
+        abort_if(!hasPermission('view trash jobs'), 403, __('You are not authorized to view trash jobs'));
         return $this->jobRepo->trashData();
     }
 
     public function restore($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('restore job'), 403, __('You are not authorized to restore job'));
         return $this->jobRepo->restore($id);
     }
 
     public function forceDelete($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('force delete job'), 403, __('You are not authorized to force delete job'));
         return $this->jobRepo->forceDelete($id);
     }
 }
