+++ app/Http/Controllers/Backend/Dashboards/Admin/RentalSpaceController.php
@@ -19,27 +19,37 @@ class RentalSpaceController extends Controller
 
     public function index()
     {
+        // apply permissions
+        abort_if(!hasPermission('view rental spaces'), 403, __('You are not authorized to view rental spaces'));
         return view('backend.dashboards.admin.pages.rental-spaces.index');
     }
 
     public function create()
     {
+        // apply permissions
+        abort_if(!hasPermission('create rental space'), 403, __('You are not authorized to create rental space'));
         return view('backend.dashboards.clinic.pages.rental-spaces.create');
     }
 
 
     public function data()
     {
+        // apply permissions
+        abort_if(!hasPermission('view rental spaces'), 403, __('You are not authorized to view rental spaces'));
         return $this->rentalSpaceRepositoryInterface->data();
     }
 
     public function store(StoreRentalSpaceRequest $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('create rental space'), 403, __('You are not authorized to create rental space'));
         return $this->rentalSpaceRepositoryInterface->store($request);
     }
 
     public function show($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('view rental space'), 403, __('You are not authorized to view rental space'));
         $rentalSpace = $this->rentalSpaceRepositoryInterface->show($id);
 
         return request()->ajax()
@@ -49,6 +59,8 @@ class RentalSpaceController extends Controller
 
     public function edit($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update rental space'), 403, __('You are not authorized to update rental space'));
         $rentalSpace = $this->rentalSpaceRepositoryInterface->show($id);
 
         return view('backend.dashboards.admin.pages.rental-spaces.edit', compact('rentalSpace'));
@@ -56,18 +68,24 @@ class RentalSpaceController extends Controller
 
     public function update(UpdateRentalSpaceRequest $request, $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update rental space'), 403, __('You are not authorized to update rental space'));
         return $this->rentalSpaceRepositoryInterface->update($request, $id);
     }
 
     public function updateStatus(Request $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('toggle rental space status'), 403, __('You are not authorized to update rental space status'));
         return $this->rentalSpaceRepositoryInterface->updateStatus($request);
     }
 
     public function destroy($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('delete rental space'), 403, __('You are not authorized to delete rental space'));
         return $this->rentalSpaceRepositoryInterface->destroy($id);
     }
 
-    
+
 }
