+++ app/Http/Controllers/Backend/Dashboards/Admin/RoleController.php
@@ -22,6 +22,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        // apply permissions
+        abort_if(!hasPermission('view roles'), 403, __('You are not authorized to view roles'));
         return $this->roleRepository->index();
     }
 
@@ -30,6 +32,8 @@ class RoleController extends Controller
      */
     public function data()
     {
+        // apply permissions
+        abort_if(!hasPermission('view roles'), 403, __('You are not authorized to view roles'));
         return $this->roleRepository->data();
     }
 
@@ -38,6 +42,8 @@ class RoleController extends Controller
      */
     public function create()
     {
+        // apply permissions
+        abort_if(!hasPermission('create role'), 403, __('You are not authorized to create roles'));
         return $this->roleRepository->create();
     }
 
@@ -46,6 +52,8 @@ class RoleController extends Controller
      */
     public function store(CreateRoleRequest $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('create role'), 403, __('You are not authorized to create roles'));
         return $this->roleRepository->store($request);
     }
 
@@ -54,6 +62,8 @@ class RoleController extends Controller
      */
     public function show(string $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('view roles'), 403, __('You are not authorized to view role'));
         return $this->roleRepository->show($id);
     }
 
@@ -62,6 +72,8 @@ class RoleController extends Controller
      */
     public function edit(string $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update role'), 403, __('You are not authorized to update role'));
         return $this->roleRepository->edit($id);
     }
 
@@ -70,6 +82,8 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, string $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update role'), 403, __('You are not authorized to update role'));
         return $this->roleRepository->update($request, $id);
     }
 
@@ -78,6 +92,8 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('delete role'), 403, __('You are not authorized to delete role'));
         return $this->roleRepository->destroy($id);
     }
 
@@ -86,6 +102,8 @@ class RoleController extends Controller
      */
     public function trash()
     {
+        // apply permissions
+        abort_if(!hasPermission('view trash roles'), 403, __('You are not authorized to view trash roles'));
         return $this->roleRepository->trash();
     }
 
@@ -94,6 +112,8 @@ class RoleController extends Controller
      */
     public function trashData()
     {
+        // apply permissions
+        abort_if(!hasPermission('view trash roles'), 403, __('You are not authorized to view trash roles'));
         return $this->roleRepository->trashData();
     }
 
@@ -102,6 +122,8 @@ class RoleController extends Controller
      */
     public function restore(string $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('restore role'), 403, __('You are not authorized to restore role'));
         return $this->roleRepository->restore($id);
     }
 
@@ -110,6 +132,8 @@ class RoleController extends Controller
      */
     public function forceDelete(string $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('force delete role'), 403, __('You are not authorized to force delete role'));
         return $this->roleRepository->forceDelete($id);
     }
 }
