+++ app/Http/Controllers/Backend/Dashboards/Admin/SpecialityController.php
@@ -18,32 +18,44 @@ class SpecialityController extends Controller
 
     public function index()
     {
+        // apply permissions
+        abort_if(!hasPermission('view specialities'), 403, __('You are not authorized to view specialities'));
         return view('backend.dashboards.admin.pages.specialities.index');
     }
 
     public function data()
     {
+        // apply permissions
+        abort_if(!hasPermission('view specialities'), 403, __('You are not authorized to view specialities'));
         return $this->repo->data();
     }
 
     public function store(StoreSpecialityRequest $request)
     {
+        // apply permissions
+        abort_if(!hasPermission('create speciality'), 403, __('You are not authorized to create specialities'));
         return $this->repo->store($request);
     }
 
     public function show($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('view specialities'), 403, __('You are not authorized to view specialities'));
         $item = $this->repo->show($id);
         return request()->ajax() ? response()->json($item) : view('backend.dashboards.admin.pages.specialities.show', compact('item'));
     }
 
     public function update(UpdateSpecialityRequest $request, $id)
     {
+        // apply permissions
+        abort_if(!hasPermission('update speciality'), 403, __('You are not authorized to update specialities'));
         return $this->repo->update($request, $id);
     }
 
     public function destroy($id)
     {
+        // apply permissions
+        abort_if(!hasPermission('delete speciality'), 403, __('You are not authorized to delete specialities'));
         return $this->repo->destroy($id);
     }
 }
